Fix error on transaction type change
When switching between the income and expense filters, if a category or account was selected, the app would throw an error. Now, every time the selected category/account is pressed, it gets reset.
