+++ lib/pages/transactions_page/widgets/accounts_tab.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:sossoldi/pages/transactions_page/widgets/transaction_type_button.dart';
 
 import '../../../constants/constants.dart';
 import '../../../constants/functions.dart';
-import '../../../constants/style.dart';
 import '../../../custom_widgets/default_container.dart';
 import '../../../model/bank_account.dart';
 import '../../../model/transaction.dart';
@@ -12,10 +12,7 @@ import '../../../providers/transactions_provider.dart';
 import 'account_list_tile.dart';
 import 'accounts_pie_chart.dart';
 
-final selectedAccountIndexProvider = StateProvider<int>((ref) => -1);
-
-final selectedTransactionTypeProvider =
-    StateProvider<TransactionType>((ref) => TransactionType.income);
+final selectedAccountIndexProvider = StateProvider.autoDispose<int>((ref) => -1);
 
 class AccountsTab extends ConsumerStatefulWidget {
   const AccountsTab({
@@ -179,86 +176,6 @@ class _AccountsTabState extends ConsumerState<AccountsTab> with Functions {
             ),
           ],
         ),
-      ),
-    );
-  }
-}
-
-/// Switch between income and expenses
-class TransactionTypeButton extends ConsumerWidget {
-  const TransactionTypeButton({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final transactionType = ref.watch(selectedTransactionTypeProvider);
-    final width = (MediaQuery.of(context).size.width - 64) * 0.5;
-    return Container(
-      height: 28,
-      decoration: const BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.all(
-          Radius.circular(5.0),
-        ),
-      ),
-      child: Stack(
-        children: [
-          AnimatedAlign(
-            alignment: Alignment(
-              (transactionType == TransactionType.income) ? -1 : 1,
-              0,
-            ),
-            curve: Curves.decelerate,
-            duration: const Duration(milliseconds: 180),
-            child: Container(
-              width: width,
-              height: 28,
-              decoration: const BoxDecoration(
-                color: blue5,
-                borderRadius: BorderRadius.all(
-                  Radius.circular(5.0),
-                ),
-              ),
-            ),
-          ),
-          GestureDetector(
-            onTap: () {
-              ref.read(selectedTransactionTypeProvider.notifier).state = TransactionType.income;
-            },
-            child: Align(
-              alignment: const Alignment(-1, 0),
-              child: Container(
-                width: width,
-                color: Colors.transparent,
-                alignment: Alignment.center,
-                child: Text(
-                  "Income",
-                  style: Theme.of(context).textTheme.bodyLarge?.copyWith(
-                      color: (transactionType == TransactionType.income) ? white : blue2),
-                ),
-              ),
-            ),
-          ),
-          GestureDetector(
-            onTap: () {
-              ref.read(selectedTransactionTypeProvider.notifier).state = TransactionType.expense;
-            },
-            child: Align(
-              alignment: const Alignment(1, 0),
-              child: Container(
-                width: width,
-                color: Colors.transparent,
-                alignment: Alignment.center,
-                child: Text(
-                  'Expenses',
-                  style: Theme.of(context).textTheme.bodyLarge?.copyWith(
-                      color: (transactionType == TransactionType.expense) ? white : blue2),
-                ),
-              ),
-            ),
-          ),
-        ],
       ),
     );
   }
