+++ lib/pages/transactions_page/widgets/categories_tab.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:sossoldi/pages/transactions_page/widgets/transaction_type_button.dart';
 
 import '../../../constants/functions.dart';
-import '../../../constants/style.dart';
 import '../../../custom_widgets/default_container.dart';
 import '../../../model/category_transaction.dart';
 import '../../../model/transaction.dart';
@@ -11,10 +11,7 @@ import '../../../providers/transactions_provider.dart';
 import 'categories_pie_chart.dart';
 import 'category_list_tile.dart';
 
-final selectedCategoryIndexProvider = StateProvider<int>((ref) => -1);
-
-final selectedTransactionTypeProvider =
-    StateProvider<TransactionType>((ref) => TransactionType.income);
+final selectedCategoryIndexProvider = StateProvider.autoDispose<int>((ref) => -1);
 
 class CategoriesTab extends ConsumerStatefulWidget {
   const CategoriesTab({
@@ -178,86 +175,6 @@ class _CategoriesTabState extends ConsumerState<CategoriesTab> with Functions {
             ),
           ],
         ),
-      ),
-    );
-  }
-}
-
-/// Switch between income and expenses
-class TransactionTypeButton extends ConsumerWidget {
-  const TransactionTypeButton({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final transactionType = ref.watch(selectedTransactionTypeProvider);
-    final width = (MediaQuery.of(context).size.width - 64) * 0.5;
-    return Container(
-      height: 28,
-      decoration: const BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.all(
-          Radius.circular(5.0),
-        ),
-      ),
-      child: Stack(
-        children: [
-          AnimatedAlign(
-            alignment: Alignment(
-              (transactionType == TransactionType.income) ? -1 : 1,
-              0,
-            ),
-            curve: Curves.decelerate,
-            duration: const Duration(milliseconds: 180),
-            child: Container(
-              width: width,
-              height: 28,
-              decoration: const BoxDecoration(
-                color: blue5,
-                borderRadius: BorderRadius.all(
-                  Radius.circular(5.0),
-                ),
-              ),
-            ),
-          ),
-          GestureDetector(
-            onTap: () {
-              ref.read(selectedTransactionTypeProvider.notifier).state = TransactionType.income;
-            },
-            child: Align(
-              alignment: const Alignment(-1, 0),
-              child: Container(
-                width: width,
-                color: Colors.transparent,
-                alignment: Alignment.center,
-                child: Text(
-                  "Income",
-                  style: Theme.of(context).textTheme.bodyLarge?.copyWith(
-                      color: (transactionType == TransactionType.income) ? white : blue2),
-                ),
-              ),
-            ),
-          ),
-          GestureDetector(
-            onTap: () {
-              ref.read(selectedTransactionTypeProvider.notifier).state = TransactionType.expense;
-            },
-            child: Align(
-              alignment: const Alignment(1, 0),
-              child: Container(
-                width: width,
-                color: Colors.transparent,
-                alignment: Alignment.center,
-                child: Text(
-                  'Expenses',
-                  style: Theme.of(context).textTheme.bodyLarge?.copyWith(
-                      color: (transactionType == TransactionType.expense) ? white : blue2),
-                ),
-              ),
-            ),
-          ),
-        ],
       ),
     );
   }
